separate generative.py and generative_event.py one being ask slash command and one being the on_message event and use compositions
